INC-18 refactor: change structure
