+++ app/auth/forgot-password/page.tsx
@@ -1,1 +1,1 @@
-export { ForgotPassword as default } from '../../../src/pages-view/forgot-password';
+export { ForgotPassword as default } from '@/pages-view/auth';
