+++ src/app/layouts/ReduxWrapper.tsx
@@ -1,5 +1,4 @@
 'use client';
-import React from 'react';
 import { Provider } from 'react-redux';
 import { store } from '@/app/store/store';
 
@@ -7,10 +6,4 @@ type Props = {
     children: React.ReactNode;
 };
 
-export const ReduxWrapper = ({ children }: Props) => {
-    return (
-        <>
-            <Provider store={store}>{children}</Provider>
-        </>
-    );
-};
+export const ReduxWrapper = ({ children }: Props) => <Provider store={store}>{children}</Provider>
