+++ src/features/auth/ui/Form.module.scss
@@ -30,11 +30,6 @@
     width: 100%;
 }
 
-.field {
-    margin-bottom: 4px;
-    display: block;
-}
-
 .legal {
     margin-bottom: 20px;
     display: flex;
