+++ src/pages-view/auth/ui/email-resending/EmailResending.module.scss
