+++ src/pages-view/auth/ui/legal-page/LegalPage.tsx
