+++ src/pages-view/auth/ui/new-password/NewPassword.tsx
