+++ src/pages-view/auth/ui/recovery/Recovery.module.scss
