+++ src/pages-view/auth/ui/registration-confirmation/RegistrationСonfirmation.module.scss
