+++ src/pages-view/auth/ui/registration-confirmation/RegistrationСonfirmation.tsx
