+++ src/pages-view/auth/ui/registration-success/RegistrationSuccess.module.scss
