--- src/pages-view/email-resending/index.ts
@@ -1,1 +1,0 @@
-export { EmailResending } from './ui/EmailResending';
