--- src/pages-view/forgot-password/index.ts
@@ -1,1 +1,0 @@
-export { ForgotPassword } from './ui/ForgotPassword';
