--- src/pages-view/legal-page/index.ts
@@ -1,1 +1,0 @@
-export { LegalPage } from './ui/LegalPage';
