--- src/pages-view/new-password/index.ts
@@ -1,1 +1,0 @@
-export { NewPassword } from './ui/NewPassword';
