--- src/pages-view/recovery/index.ts
@@ -1,1 +1,0 @@
-export { Recovery } from './ui/Recovery';
