--- src/pages-view/registration-confirmation/index.ts
@@ -1,1 +1,0 @@
-export { RegistrationConfirmation } from './ui/RegistrationСonfirmation';
