--- src/pages-view/registration-success/index.ts
@@ -1,1 +1,0 @@
-export { RegistrationSuccess } from './ui/RegistrationSuccess';
