--- src/pages-view/sign-in/index.ts
@@ -1,1 +1,0 @@
-export { SignIn } from './ui/SignIn';
