--- src/pages-view/sign-up/index.ts
@@ -1,1 +1,0 @@
-export { SignUp } from './ui/SignUp';
